+++ changelog.md
@@ -2,6 +2,7 @@
 ## v1.12.4-alpha.2
 - Now up to date with the removed features from v1.12.3
 - Readded mac support
+- Fixed remove player from main menu not removing ship fire (#15)
 ## v1.12.4-alpha.1
 - Geode 3.0.0 alpha build
 - No mac / betterinfo support